backend/kg/requirements.txt
# Copyright 2025, Battelle Energy Alliance, LLC, ALL RIGHTS RESERVED

alembic==1.15.2
annotated-types==0.7.0
anyio==4.7.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.6
h11==0.16.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.2.0
packaging==24.2
pathspec==0.12.1
pgvector==0.3.6
platformdirs==4.3.6
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
requests==2.32.4
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.47.2
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.32.1
watchfiles==1.0.3
websockets==14.1
